requirements.txt
kafka-python==2.0.1
numpy==1.18.2
opencv-python==4.2.0.34